refactor: NewAESCommand has been changed to AESCommand
