+++ src/LionBundle/Commands/Lion/New/AESCommand.php
@@ -2,21 +2,21 @@
 
 declare(strict_types=1);
 
-namespace Lion\Bundle\Commands\Lion\AES;
+namespace Lion\Bundle\Commands\Lion\New;
 
 use Lion\Bundle\Traits\AES;
 use Lion\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class NewAESCommand extends Command
+class AESCommand extends Command
 {
     use AES;
 
     protected function configure(): void
     {
         $this
-            ->setName('aes:new')
+            ->setName('new:aes')
             ->setDescription('Command to create KEY and IV keys for AES');
     }
